Crate automerge_test

Macros

assert_doc

This macro makes it easy to make assertions about a document. It is called with two arguments, the first is a reference to an automerge::Automerge, the second is an instance of RealizedObject<ExportableOpId>.

assert_obj

Like assert_doc except that you can specify an object ID and property to select subsections of the document.

list

Construct RealizedObject::Sequence. This macro represents a sequence of values

map

Construct RealizedObject::Map. This macro takes a nested set of curl braces. The outer set is the keys of the map, the inner set is the set of values for that key:

Structs

ExportedOpId

Enums

OrdScalarValue

RealizedObject

A RealizedObject is a representation of all the current values in a document - including conflicts.

Functions

mk_counter

new_doc

new_doc_with_actor

pretty_panic

pretty_print

Pretty print the contents of a document

realize

realize_obj

realize_prop

sorted_actors

Returns two actor IDs, the first considered to be ordered before the second